+++ src/package.json
@@ -41,13 +41,13 @@
     "lint": "eslint --ext .ts .",
     "test:e2e": "./node_modules/.bin/jest --testMatch '**/e2e/?(*.)test.ts' --passWithNoTests",
     "test:infra": "./node_modules/.bin/jest --testMatch '**/infra/?(*.)test.ts' ",
-    "api": "JEST_JUNIT_OUTPUT_NAME=api-report.xml ./node_modules/.bin/jest --runInBand --testPathPattern=tests/api/",
+    "api": "JEST_JUNIT_OUTPUT_NAME=api-report.xml ./node_modules/.bin/jest --runInBand --testPathPattern=tests/api/backend",
     "test:api": "npm run compile && npm run api",
-    "yoti": "JEST_JUNIT_OUTPUT_NAME=api-report.xml ./node_modules/.bin/jest --runInBand --testPathPattern=tests/YotiStub.test.ts",
+    "yoti": "JEST_JUNIT_OUTPUT_NAME=api-report.xml ./node_modules/.bin/jest --runInBand --testPathPattern=tests/api/stub/YotiStub.test.ts",
     "test:yoti": "npm run compile && npm run yoti",
-    "postoffice": "JEST_JUNIT_OUTPUT_NAME=api-report.xml ./node_modules/.bin/jest --runInBand --testPathPattern=tests/api/PostOfficeStub.test.ts",
+    "postoffice": "JEST_JUNIT_OUTPUT_NAME=api-report.xml ./node_modules/.bin/jest --runInBand --testPathPattern=tests/api/stub/PostOfficeStub.test.ts",
     "test:postoffice": "npm run compile && npm run postoffice",
-    "govnotify": "JEST_JUNIT_OUTPUT_NAME=api-report.xml ./node_modules/.bin/jest --runInBand --testPathPattern=tests/api/GovNotifyStub.test.ts",
+    "govnotify": "JEST_JUNIT_OUTPUT_NAME=api-report.xml ./node_modules/.bin/jest --runInBand --testPathPattern=tests/api/stub/GovNotifyStub.test.ts",
     "test:govnotify": "npm run compile && npm run govnotify",
     "test:pii": "bash ./check-logs.sh",
     "kill:dynamodblocal": "cd tests/contract; ./scripts/kill-dynamodb-local.sh",
