+++ src/tests/api/stub/YotiStub.test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable max-lines-per-function */
-import yotiRequestData from "./data/yotiSessionsPayloadValid.json";
+import yotiRequestData from "../../data/yotiSessionsPayloadValid.json";
 import { 
 	postYotiSession, 
 	getYotiSessionsConfiguration,
